Rename launchers and add mode selector
