+++ web.py
@@ -44,21 +44,25 @@ def voice_to_text() -> str:
     return transcribe_whisper()
 
 
-with gr.Blocks() as demo:
-    gr.Markdown("# Nira Chat")
-    chatbot = gr.Chatbot(type="messages")
-    with gr.Row():
-        msg = gr.Textbox(label="Message", scale=4)
-        submit = gr.Button("Submit", scale=1)
-        voice_btn = gr.Button("Voice", scale=1)
-    with gr.Row():
-        speak_toggle = gr.Checkbox(False, label="Speak")
-        clear = gr.Button("Clear")
+def main() -> None:
+    with gr.Blocks() as demo:
+        gr.Markdown("# Nira Chat")
+        chatbot = gr.Chatbot(type="messages")
+        with gr.Row():
+            msg = gr.Textbox(label="Message", scale=4)
+            submit = gr.Button("Submit", scale=1)
+            voice_btn = gr.Button("Voice", scale=1)
+        with gr.Row():
+            speak_toggle = gr.Checkbox(False, label="Speak")
+            clear = gr.Button("Clear")
 
-    submit.click(chat, [msg, chatbot, speak_toggle], [msg, chatbot])
-    msg.submit(chat, [msg, chatbot, speak_toggle], [msg, chatbot])
-    voice_btn.click(voice_to_text, None, msg)
-    clear.click(lambda: [], None, chatbot, queue=False)
+        submit.click(chat, [msg, chatbot, speak_toggle], [msg, chatbot])
+        msg.submit(chat, [msg, chatbot, speak_toggle], [msg, chatbot])
+        voice_btn.click(voice_to_text, None, msg)
+        clear.click(lambda: [], None, chatbot, queue=False)
+
+    demo.launch()
 
 
-demo.launch()
+if __name__ == "__main__":
+    main()
